Simplifying retrieval of appCapabilities for an org unit - extracting categorizeCostsIntoBuckets into util file
#25
