+++ waltz-ng/client/org-units/services/org-unit-view-data.js
@@ -14,6 +14,7 @@ import _ from 'lodash';
 
 import RatedFlowsData from '../../data-flow/RatedFlowsData';
 import { aggregatePeopleInvolvements } from '../../involvement/involvement-utils';
+import { categorizeCostsIntoBuckets } from '../../asset-cost/services/asset-cost-utilities';
 
 
 function prepareFlowData(flows, apps) {
@@ -103,40 +104,6 @@ function extractPrimaryAppIds(appCapabilities) {
         .value();
 }
 
-
-function categorizeCostsIntoBuckets(costs) {
-
-    // bucketPredicate:: bucket -> amount -> bool
-    const bucketPredicate = b => a => (a >= b.min) && (a < b.max);
-
-    const buckets = [
-        { min: 0, max: 1000, name: '€0 < 1K', idx: 0, size: 0},
-        { min: 1000, max: 5000, name: '€1K < 5K', idx: 1, size: 0},
-        { min: 5000, max: 10000, name: '€5K < 10K', idx: 2, size: 0},
-        { min: 10000, max: 50000, name: '€10K < 50K', idx: 3, size: 0},
-        { min: 50000, max: 100000, name: '€50K < 100K', idx: 4, size: 0},
-        { min: 100000, max: 500000, name: '€100K < 500K', idx: 5, size: 0},
-        { min: 500000, max: 1000000, name: '€500K < 1M', idx: 6, size: 0},
-        { min: 1000000, max: Number.MAX_VALUE, name: '€1M +', idx: 7, size: 0}
-    ];
-
-    const findBucket = (c) => {
-        return _.find(buckets, b => bucketPredicate(b)(c));
-    };
-
-
-    _.each(costs, c => {
-        const bucket = findBucket(c.cost.amount);
-        if (bucket) {
-            bucket.size++;
-        } else {
-            console.log('failed to find bucket for ', c);
-        }
-    });
-
-    return buckets;
-
-}
 
 function service(appStore,
                  appCapabilityStore,
@@ -291,5 +258,6 @@ service.$inject = [
     'ComplexityStore',
     '$q'
 ];
+
 
 export default service;
